FACES-2965 Develop tests for the Multimedia category of the Liferay Faces Showcase
