+++ jsf-showcase-webapp/src/main/webapp/WEB-INF/component/h/graphicimage/general/graphicImage.xhtml
@@ -36,7 +36,7 @@
 			<showcase:example description="#{i18n['h-graphicimage-general-example-4']}" number="4">
 				<h:graphicImage usemap="#logoMapName" value="#{resource['images:jsf-logo-small.png']}" />
 				<map name="logoMapName">
-					<area alt="JSR 362" coords="1,1,110,136" href="https://jcp.org/en/jsr/detail?id=344" shape="rect" target="_blank"
+					<area alt="JSR 362" coords="1,1,110,136" href="https://jcp.org/en/jsr/detail?id=362" shape="rect" target="_blank"
 						title="JSR 362" />
 					<area alt="JSR 378" coords="110,1,190,136" href="https://www.jcp.org/en/jsr/detail?id=378" shape="rect"
 						target="_blank" title="JSR 378" />
